add SepReader.IO.Async/Sync with initial "duplicates"
